decimal -> integer

+++ src/migrations/20240924141755-create-goods-receipt-detail.js
@@ -16,7 +16,7 @@ module.exports = {
         type: Sequelize.INTEGER
       },
       cost: {
-        type: Sequelize.DECIMAL(20,2)
+        type: Sequelize.INTEGER
       },
       goodsReceiptId: {
         type: Sequelize.INTEGER
