+++ src/models/invoice.js
@@ -17,7 +17,7 @@ module.exports = (sequelize, DataTypes) => {
     }
   }
   Invoice.init({
-    totalCost: DataTypes.DECIMAL(20,2),
+    totalCost: DataTypes.INTEGER,
     status: DataTypes.ENUM('paid', 'canceled','pending'),
     paymentMethod: DataTypes.ENUM('Cash', 'QR code'),
   }, {
